Remove "pilot" namespace for registrations
Unnecessary nesting.
