+++ app/controllers/registrations_controller.rb
@@ -1,0 +1,16 @@
+class RegistrationsController < ApplicationController
+  skip_before_action :authenticate_user!
+  before_action :set_school
+
+  layout "registration"
+
+  def new
+    @parent_interest_form = Registration.new
+  end
+
+  private
+
+  def set_school
+    @school = Location.find(params[:school_id])
+  end
+end
