+++ app/views/registrations/new.html.erb
