+++ config/routes.rb
@@ -93,12 +93,12 @@ Rails.application.routes.draw do
     post "setup-offline", to: "offline_passwords#create", on: :member
   end
 
-  resources :schools, only: [:show] do
-    resources :registrations,
-              only: %i[new create],
-              controller: "pilot/registrations"
-  end
+  resources :schools, only: [:show]
   resources :consent_forms, only: [:show]
+
+  resource :registration,
+           path: "/schools/:school_id/registration",
+           only: %i[new create update]
 
   scope via: :all do
     get "/404", to: "errors#not_found"
